fix(uint256): Remove temporary values in Dec loop
Modified the implementation of the `Dec` function to reuse temporary variables, similar to the approach applied to `fromDecimal`.

After the fix, the `Dec` function saves 1MB of memory over 200 iterations. Additionally, since the `ToString` function internally calls the `Dec` function, the same improvement can be observed.

| Function | Before Optimization | After Optimization  | Improvement | Improvement Rate |
|----------|------------|------------|------------|-------|
| Dec      | 56,978,184 | 55,972,384 | -1,005,800 | -1.8% |
| ToString | 57,306,184 | 56,300,384 | -1,005,800 | -1.8% |

Key changes are as follows:
- []byte("000...") → [98]byte array
- make([]byte, 0, 19) → [19]byte array
- Removed new(Uint) and use value copy instead